+++ configserver/src/main/java/com/yahoo/vespa/config/server/http/v2/SessionPrepareHandler.java
@@ -34,7 +34,7 @@ import java.util.logging.Logger;
  * @author hmusum
  * @since 5.1.29
  */
-// TODO: Move business logic out of the HTTP layer and delegate to a Deployer
+// TODO: Move business logic out of the HTTP layer and delegate to a ApplicationRepository
 public class SessionPrepareHandler extends SessionHandler {
 
     private static final Logger log = Logger.getLogger(SessionPrepareHandler.class.getName());
