+++ configserver/src/main/resources/configserver-app/services.xml
@@ -20,7 +20,7 @@
     <component id="com.yahoo.vespa.config.server.Tenants" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.application.PermanentApplicationPackage" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.HostRegistries" bundle="configserver" />
-    <component id="com.yahoo.vespa.config.server.deploy.Deployer" bundle="configserver" />
+    <component id="com.yahoo.vespa.config.server.ApplicationRepository" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.version.VersionState" bundle="configserver" />
     <component id="com.yahoo.vespa.config.server.zookeeper.ConfigCurator" bundle="configserver" />
     <component id="com.yahoo.container.jdisc.metric.state.StateMetricConsumerFactory" bundle="container-disc" />
